requirements.txt
bidict==0.23.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
databricks-connect==15.1.0
databricks-sdk==0.29.0
Faker==26.0.0
Flask==3.0.3
Flask-SocketIO==5.3.6
google-auth==2.32.0
googleapis-common-protos==1.63.2
grpcio==1.64.1
grpcio-status==1.64.1
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
packaging==24.1
pandas==2.2.2
protobuf==5.27.2
py4j==0.10.9.7
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
python-dateutil==2.9.0.post0
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
requests==2.32.3
rsa==4.9
simple-websocket==1.0.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
